+++ model/util-createdb.js
@@ -40,6 +40,7 @@ const PublicPollInvite = require("./publicpollinvite");
 const ContactSuppression = require("./contactsuppression");
 const TrustProfile = require("./trustprofile");
 const AbuseAudit = require("./abuseaudit");
+const RateLimit = require("./ratelimit");
 const TerminologyUpdate = require("../updates/update-0019");
 
 let createdFreshSchema = false;